Refactoring of Feature Executions
